refactor: use hooks on assets coin
